# frontend/src/api/generated/openapi.schema.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "CUHKSZ Group Discussion API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/system/demo-scenarios/{scenario_id}/run": {
      "post": {
        "tags": [
          "system"
        ],
        "summary": "Run Demo Scenario Ep",
        "operationId": "run_demo_scenario_ep_system_demo_scenarios__scenario_id__run_post",
        "parameters": [
          {
            "name": "scenario_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Scenario Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DemoScenarioRunRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DemoScenarioResultResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/system/e2e-scenarios/{scenario_id}/run": {
      "post": {
        "tags": [
          "stability"
        ],
        "summary": "Run E2E",
        "operationId": "run_e2e_system_e2e_scenarios__scenario_id__run_post",
        "parameters": [
          {
            "name": "scenario_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Scenario Id"
            }
          },
          {
            "name": "profile_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Profile Id"
            }
          },
          {
            "name": "snapshot_id",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "dev_snapshot_v2",
              "title": "Snapshot Id"
            }
          },
          {
            "name": "topic_id",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "tc-campus-ai",
              "title": "Topic Id"
            }
          },
          {
            "name": "provider_name",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "mock",
              "title": "Provider Name"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/E2ERunResultResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/sessions/{session_id}/reviews": {
      "get": {
        "tags": [
          "reviews"
        ],
        "summary": "Session Reviews",
        "operationId": "session_reviews_sessions__session_id__reviews_get",
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Session Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "additionalProperties": true
                  },
                  "title": "Response Session Reviews Sessions  Session Id  Reviews Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/runtime-reviews": {
      "get": {
        "tags": [
          "runtime-reviews"
        ],
        "summary": "List Runtime Reviews",
        "operationId": "list_runtime_reviews_runtime_reviews_get",
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "pending",
              "title": "Status"
            }
          },
          {
            "name": "session_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Session Id"
            }
          },
          {
            "name": "topic_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Topic Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RuntimeReviewSummaryResponse"
                  },
                  "title": "Response List Runtime Reviews Runtime Reviews Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/runtime-reviews/{review_id}": {
      "get": {
        "tags": [
          "runtime-reviews"
        ],
        "summary": "Get Runtime Review",
        "operationId": "get_runtime_review_runtime_reviews__review_id__get",
        "parameters": [
          {
            "name": "review_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Review Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RuntimeReviewDetailResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/runtime-reviews/{review_id}/approve": {
      "post": {
        "tags": [
          "runtime-reviews"
        ],
        "summary": "Approve Runtime Review",
        "operationId": "approve_runtime_review_runtime_reviews__review_id__approve_post",
        "parameters": [
          {
            "name": "review_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Review Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ApproveRuntimeReviewRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RuntimeReviewActionResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/runtime-reviews/{review_id}/reject": {
      "post": {
        "tags": [
          "runtime-reviews"
        ],
        "summary": "Reject Runtime Review",
        "operationId": "reject_runtime_review_runtime_reviews__review_id__reject_post",
        "parameters": [
          {
            "name": "review_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Review Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RejectRuntimeReviewRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RuntimeReviewActionResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/runtime-reviews/{review_id}/resume": {
      "post": {
        "tags": [
          "runtime-reviews"
        ],
        "summary": "Resume Runtime Review",
        "operationId": "resume_runtime_review_runtime_reviews__review_id__resume_post",
        "parameters": [
          {
            "name": "review_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Review Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResumeRuntimeReviewRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Resume Runtime Review Runtime Reviews  Review Id  Resume Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/runtime-reviews/{review_id}/apply-edited-draft": {
      "post": {
        "tags": [
          "runtime-reviews"
        ],
        "summary": "Apply Edited Draft",
        "operationId": "apply_edited_draft_runtime_reviews__review_id__apply_edited_draft_post",
        "parameters": [
          {
            "name": "review_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Review Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ApplyEditedDraftRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Apply Edited Draft Runtime Reviews  Review Id  Apply Edited Draft Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ApplyEditedDraftRequest": {
        "properties": {
          "edited_text": {
            "type": "string",
            "title": "Edited Text"
          },
          "expected_version": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expected Version"
          },
          "updated_by": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Updated By"
          },
          "note": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Note"
          },
          "resume_after_apply": {
            "type": "boolean",
            "title": "Resume After Apply",
            "default": false
          },
          "additional_steps": {
            "type": "integer",
            "title": "Additional Steps",
            "default": 1
          }
        },
        "type": "object",
        "required": [
          "edited_text"
        ],
        "title": "ApplyEditedDraftRequest"
      },
      "ApproveRuntimeReviewRequest": {
        "properties": {
          "action": {
            "type": "string",
            "title": "Action",
            "default": "approve"
          },
          "expected_version": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expected Version"
          },
          "updated_by": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Updated By"
          },
          "payload": {
            "additionalProperties": true,
            "type": "object",
            "title": "Payload"
          }
        },
        "type": "object",
        "title": "ApproveRuntimeReviewRequest"
      },
...
      "DemoScenarioResultResponse": {
        "properties": {
          "result": {
            "additionalProperties": true,
            "type": "object",
            "title": "Result"
          }
        },
        "type": "object",
        "required": [
          "result"
        ],
        "title": "DemoScenarioResultResponse"
      },
      "DemoScenarioRunRequest": {
        "properties": {
          "profile_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Profile Id"
          },
          "snapshot_id": {
            "type": "string",
            "title": "Snapshot Id",
            "default": "dev_snapshot_v2"
          },
          "topic_id": {
            "type": "string",
            "title": "Topic Id",
            "default": "tc-campus-ai"
          },
          "provider_name": {
            "type": "string",
            "title": "Provider Name",
            "default": "mock"
          }
        },
        "type": "object",
        "title": "DemoScenarioRunRequest"
      },
...
      "E2ERunResultResponse": {
        "properties": {
          "result": {
            "additionalProperties": true,
            "type": "object",
            "title": "Result"
          }
        },
        "type": "object",
        "required": [
          "result"
        ],
        "title": "E2ERunResultResponse"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "RejectRuntimeReviewRequest": {
        "properties": {
          "reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reason"
          },
          "expected_version": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expected Version"
          },
          "updated_by": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Updated By"
          }
        },
        "type": "object",
        "title": "RejectRuntimeReviewRequest"
      },
...
      "ResumeRuntimeReviewRequest": {
        "properties": {
          "additional_steps": {
            "type": "integer",
            "title": "Additional Steps",
            "default": 1
          }
        },
        "type": "object",
        "title": "ResumeRuntimeReviewRequest"
      },
...
      "RuntimeReviewActionResponse": {
        "properties": {
          "review": {
            "$ref": "#/components/schemas/RuntimeReviewDetailResponse"
          },
          "result": {
            "additionalProperties": true,
            "type": "object",
            "title": "Result"
          }
        },
        "type": "object",
        "required": [
          "review"
        ],
        "title": "RuntimeReviewActionResponse"
      },
      "RuntimeReviewDetailResponse": {
        "properties": {
          "review_id": {
            "type": "string",
            "title": "Review Id"
          },
          "session_id": {
            "type": "string",
            "title": "Session Id"
          },
          "run_id": {
            "type": "string",
            "title": "Run Id"
          },
          "topic_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Topic Id"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "interrupt_reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Interrupt Reason"
          },
          "quality_flags": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Quality Flags"
          },
          "version": {
            "type": "integer",
            "title": "Version"
          },
          "allowed_actions": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Allowed Actions"
          },
          "updated_at": {
            "type": "string",
            "title": "Updated At"
          },
          "checkpoint_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Checkpoint Id"
          },
          "reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reason"
          },
          "draft_reply_text": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Draft Reply Text"
          },
          "draft_reply_summary": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Draft Reply Summary"
          },
          "review_decision_payload": {
            "additionalProperties": true,
            "type": "object",
            "title": "Review Decision Payload"
          },
          "notes": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Notes"
          },
          "resolved_by": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Resolved By"
          },
          "updated_by": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Updated By"
          },
          "metadata": {
            "additionalProperties": true,
            "type": "object",
            "title": "Metadata"
          },
          "created_at": {
            "type": "string",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "review_id",
          "session_id",
          "run_id",
          "status",
          "version",
          "updated_at",
          "created_at"
        ],
        "title": "RuntimeReviewDetailResponse"
      },
...
      "RuntimeReviewSummaryResponse": {
        "properties": {
          "review_id": {
            "type": "string",
            "title": "Review Id"
          },
          "session_id": {
            "type": "string",
            "title": "Session Id"
          },
          "run_id": {
            "type": "string",
            "title": "Run Id"
          },
          "topic_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Topic Id"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "interrupt_reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Interrupt Reason"
          },
          "quality_flags": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Quality Flags"
          },
          "version": {
            "type": "integer",
            "title": "Version"
          },
          "allowed_actions": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Allowed Actions"
          },
          "updated_at": {
            "type": "string",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "review_id",
          "session_id",
          "run_id",
          "status",
          "version",
          "updated_at"
        ],
        "title": "RuntimeReviewSummaryResponse"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}